Move index to root and remove redirection to questions, add test, restructure css
