+++ src/components/feedbacks/FeedbackChannel.js
@@ -74,7 +74,7 @@ const FeedbackChannel = ({
     )}
 
     <style jsx>{`
-      @import 'src/_theme';
+      @import 'src/theme';
 
       .feedbackChannel {
         display: flex;
