+++ src/pages/student/session.js
@@ -162,7 +162,7 @@ const Session = ({
         </div>
 
         <style jsx>{`
-          @import 'src/_theme';
+          @import 'src/theme';
 
           .student {
             display: flex;
