+++ src/pages/user/login.js
@@ -34,7 +34,7 @@ const Login = ({ intl, error, handleSubmit }) => (
       <LoginForm intl={intl} onSubmit={handleSubmit} />
 
       <style jsx>{`
-        @import 'src/_theme';
+        @import 'src/theme';
 
         .login {
           padding: 1rem;
